@@ -6,59 +6,59 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "test": "ng test --colors=true --watch=true --progress=true --sourcemaps=false ",
-    "pree2e": "webdriver-manager update --standalone false --gecko false",
-    "e2e": "protractor"
+    "build": "ng build",
+    "test": "jest",
+    "lint": "ng lint",
+    "cypress": "node tools/run-cypress.js",
+    "e2e": "node tools/run-e2e.js"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.3",
-    "@angular/cdk": "^6.1.0",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.1.0",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/platform-server": "^6.0.3",
-    "@angular/router": "^6.0.3",
-    "@ngrx/core": "^1.2.0",
-    "@ngrx/effects": "^6.0.1",
-    "@ngrx/router-store": "^6.0.1",
-    "@ngrx/store": "^6.0.1",
-    "@ngrx/store-devtools": "^6.0.1",
-    "@ngrx/store-log-monitor": "^3.0.2",
-    "angular2-websocket": "^0.9.6",
-    "core-js": "^2.5.6",
-    "font-awesome": "^4.7.0",
-    "ngrx-store-localstorage": "^5.0.0",
-    "primeng": "^6.0.0-alpha.1",
+    "@angular/animations": "^18.2.14",
+    "@angular/cdk": "^18.2.8",
+    "@angular/common": "^18.2.14",
+    "@angular/compiler": "^18.2.14",
+    "@angular/core": "^18.2.14",
+    "@angular/forms": "^18.2.14",
+    "@angular/material": "^18.2.8",
+    "@angular/platform-browser": "^18.2.14",
+    "@angular/platform-browser-dynamic": "^18.2.14",
+    "@angular/platform-server": "^18.2.14",
+    "@angular/router": "^18.2.14",
+    "@ngrx/effects": "^18.1.0",
+    "@ngrx/router-store": "^18.1.0",
+    "@ngrx/store": "^18.1.0",
+    "@ngrx/store-devtools": "^18.1.0",
+    "core-js": "^3.37.1",
     "raven-js": "^3.25.2",
     "reselect": "^3.0.1",
-    "rxjs": "^6.2.0",
+    "rxjs": "^7.8.1",
     "ts-helpers": "^1.1.2",
-    "zone.js": "^0.8.26"
+    "tslib": "^2.6.2",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.5",
-    "@angular/cli": "^6.0.5",
-    "@angular/compiler-cli": "^6.0.3",
-    "@types/jasmine": "2.8.7",
-    "@types/node": "^10.1.2",
-    "codelyzer": "^4.3.0",
-    "jasmine-core": "3.1.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.2",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.1.2",
-    "karma-remap-istanbul": "^0.6.0",
-    "ngrx-store-freeze": "^0.2.3",
-    "protractor": "~5.3.2",
-    "ts-node": "6.0.5",
-    "tslint": "^5.10.0",
-    "typescript": "<2.8"
+    "@angular-devkit/build-angular": "^18.2.21",
+    "@angular/cli": "^18.2.21",
+    "@angular/compiler-cli": "^18.2.14",
+    "@angular-eslint/builder": "^18.3.0",
+    "@angular-eslint/eslint-plugin": "^18.3.0",
+    "@angular-eslint/eslint-plugin-template": "^18.3.0",
+    "@angular-eslint/schematics": "^18.3.0",
+    "@angular-eslint/template-parser": "^18.3.0",
+    "@angular-builders/jest": "^20.0.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.11.30",
+    "cypress": "^13.15.0",
+    "http-server": "^14.1.1",
+    "eslint": "^8.57.0",
+    "eslint-plugin-rxjs": "^5.0.3",
+    "jest": "^29.7.0",
+    "jest-preset-angular": "^14.6.2",
+    "@typescript-eslint/eslint-plugin": "^8.11.0",
+    "@typescript-eslint/parser": "^8.11.0",
+    "@typescript-eslint/utils": "^8.11.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "~5.4.5"
   }
 }
