@@ -1,12 +1,8 @@
 import { Injectable } from '@angular/core';
-import {
-  HttpClient,
-  HttpHeaders,
-  HttpParams,
-  HttpResponse,
-} from '@angular/common/http';
+import { HttpClient, HttpHeaders, HttpParams, HttpResponse } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import { switchMap, take, timeout } from 'rxjs/operators';
 
 import { Store } from '@ngrx/store';
 
@@ -31,58 +27,61 @@ export class RestClient {
   }
 
   get(schemaName: string, endpoint: string, params?: HttpParams): Observable<Object> {
-    return this.store.take(1).switchMap(state => {
+    return this.store.pipe(
+      take(1),
+      switchMap(state => {
         let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
         headers = headers.set('prefer', 'return=representation');
         headers = headers.set('prefer', 'count=exact');
-        return this.http.get(RestClient.createEndpoint(schemaName, endpoint),
-          {
-            headers: headers,
-            observe: 'response',
-            params: params,
-          }).timeout(this._timeout);
-      },
-    );
-  };
-
-  post(schemaName: string, endpoint: string, data: any): Observable<Object> {
-    return this.store.take(1).switchMap(state => {
-        let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
-        headers = headers.set('prefer', 'return=representation');
-        return this.http.post(RestClient.createEndpoint(schemaName, endpoint), data,
-          {
-            headers: headers,
-            observe: 'response',
-          }).timeout(this._timeout);
-      },
-    );
-  };
-
-  delete(schemaName: string, endpoint: string, id: string): Observable<Object> {
-    return this.store.take(1).switchMap(state => {
-      const params: HttpParams = new HttpParams();
-      let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
-      headers = headers.set('prefer', 'return=minimal');
-      params.set('id', 'eq.' + id);
-      return this.http.delete(RestClient.createEndpoint(schemaName, endpoint), {
-        headers: headers,
-        observe: 'response',
-        params: params,
-      }).timeout(this._timeout);
-    });
-  }
-
-  patch(schemaName: string, endpoint: string, data: any, params: HttpParams): Observable<Object> {
-    return this.store.take(1).switchMap(state => {
-      let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
-      headers = headers.set('prefer', 'return=representation');
-      return this.http.patch(RestClient.createEndpoint(schemaName, endpoint), data,
-        {
+        return this.http.get(RestClient.createEndpoint(schemaName, endpoint), {
           headers: headers,
           observe: 'response',
           params: params,
-        }).timeout(this._timeout);
-    });
+        }).pipe(timeout(this._timeout));
+      }));
+  };
+
+  post(schemaName: string, endpoint: string, data: any): Observable<Object> {
+    return this.store.pipe(
+      take(1),
+      switchMap(state => {
+        let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
+        headers = headers.set('prefer', 'return=representation');
+        return this.http.post(RestClient.createEndpoint(schemaName, endpoint), data, {
+          headers: headers,
+          observe: 'response',
+        }).pipe(timeout(this._timeout));
+      }));
+  };
+
+  delete(schemaName: string, endpoint: string, id: string): Observable<Object> {
+    return this.store.pipe(
+      take(1),
+      switchMap(state => {
+      let params: HttpParams = new HttpParams();
+      let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
+      headers = headers.set('prefer', 'return=minimal');
+      params = params.set('id', 'eq.' + id);
+      return this.http.delete(RestClient.createEndpoint(schemaName, endpoint), {
+        headers: headers,
+        observe: 'response',
+        params: params
+      }).pipe(timeout(this._timeout));
+    }));
+  }
+
+  patch(schemaName: string, endpoint: string, data: any, params: HttpParams): Observable<Object> {
+    return this.store.pipe(
+      take(1),
+      switchMap(state => {
+      let headers = RestClient.createAuthorizationHeader(new HttpHeaders(), state.auth.token);
+      headers = headers.set('prefer', 'return=representation');
+      return this.http.patch(RestClient.createEndpoint(schemaName, endpoint), data, {
+        headers: headers,
+        observe: 'response',
+        params: params
+      }).pipe(timeout(this._timeout));
+    }));
   }
 
   constructor(private http: HttpClient,
