@@ -1,31 +1,32 @@
 import { Component, Input } from '@angular/core';
-import {FormGroup} from '@angular/forms';
+import { FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'app-dynamic-form-element',
   template: `
-    <div class="ui-g" [formGroup]="formGroup">
-      <div class="ui-g-6">
-        <label>
-          {{formFieldLabel}}
-        </label>
-      </div>
-      <div class="ui-g-6">
-        <input *ngIf="formFieldName !== 'password'"
-                [formControlName]="formFieldName"
-                [id]="formFieldName"
-                pInputText
-        />
-
-        <input *ngIf="formFieldName === 'password'"
-               [formControlName]="formFieldName"
-               [id]="formFieldName"
-               type="password"
-               pPassword
-        />
-      </div>
-    </div>
-  `
+    <mat-form-field appearance="outline" class="form-field" [formGroup]="formGroup">
+      <mat-label>{{ formFieldLabel }}</mat-label>
+      <input
+        *ngIf="formFieldName !== 'password'"
+        matInput
+        [formControlName]="formFieldName"
+        [id]="formFieldName"
+        type="text"
+      />
+      <input
+        *ngIf="formFieldName === 'password'"
+        matInput
+        [formControlName]="formFieldName"
+        [id]="formFieldName"
+        type="password"
+      />
+    </mat-form-field>
+  `,
+  styles: [`
+    .form-field {
+      width: 100%;
+    }
+  `]
 })
 export class FormElementComponent {
   @Input() formFieldName: string;
