@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 
-import { Actions, Effect } from '@ngrx/effects';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
 
-import { of } from 'rxjs/observable/of';
+import { of } from 'rxjs';
+import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
 
 import {
   ReceiveFormFieldSettingsAction,
@@ -16,48 +17,50 @@ import { RouteParams } from '../router/router.models';
 @Injectable()
 export class FormEffects {
 
-  @Effect()
-  selectForm$ = this.actions$
-    .ofType(SELECT_FORM)
-    .map((action: SelectFormAction) => action.payload)
-    .mergeMap((routeParams: RouteParams) => {
+  
+  selectForm$ = createEffect(() => this.actions$.pipe(
+    ofType(SELECT_FORM),
+    map((action: SelectFormAction) => action.payload),
+    mergeMap((routeParams: RouteParams) => {
       return [
         new GetFormSettingsAction(routeParams),
         new GetFormFieldSettingsAction(routeParams),
       ];
-    });
+    })));
 
-  @Effect()
-  getFormSettings$ = this.actions$
-    .ofType(GET_FORM_SETTINGS)
-    .switchMap((action: GetFormSettingsAction) => {
+  
+  getFormSettings$ = createEffect(() => this.actions$.pipe(
+    ofType(GET_FORM_SETTINGS),
+    switchMap((action: GetFormSettingsAction) => {
       return this.formService.get_form_settings(action.payload.selectedSchemaName,
-        action.payload.selectedObjectName)
-        .mergeMap((response: any) => {
+        action.payload.selectedObjectName).pipe(
+        mergeMap((response: any) => {
           return [
             new ReceiveFormSettingsAction(response.body),
           ];
-        })
-        .catch(error => {
+        }),
+        catchError(error => {
           return of(new ReceiveFormSettingsAction(error));
-        })
-    });
+        }),
+      )
+    })));
 
-  @Effect()
-  getFormFieldSettings$ = this.actions$
-    .ofType(GET_FORM_FIELD_SETTINGS)
-    .switchMap((action: GetFormFieldSettingsAction) => {
+  
+  getFormFieldSettings$ = createEffect(() => this.actions$.pipe(
+    ofType(GET_FORM_FIELD_SETTINGS),
+    switchMap((action: GetFormFieldSettingsAction) => {
       return this.formService.get_form_field_settings(action.payload.selectedSchemaName,
-        action.payload.selectedObjectName)
-        .mergeMap((response: any) => {
+        action.payload.selectedObjectName).pipe(
+        mergeMap((response: any) => {
           return [
             new ReceiveFormFieldSettingsAction(response.body),
           ];
-        })
-        .catch(error => {
+        }),
+        catchError(error => {
           return of(new ReceiveFormFieldSettingsAction(error));
-        })
-    });
+        }),
+      )
+    })));
 
 
   constructor(private actions$: Actions,
