@@ -1,11 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
-
-import { ButtonModule } from 'primeng/components/button/button';
-import { FieldsetModule } from 'primeng/components/fieldset/fieldset';
-import { InputTextModule } from 'primeng/components/inputtext/inputtext';
-import { PasswordModule } from 'primeng/components/password/password';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
 
 import { FormContainer } from './form.container';
 import { FormComponent } from './form.component';
@@ -14,11 +13,11 @@ import { FormService } from './form.service';
 
 @NgModule({
   imports: [
-    ButtonModule,
     CommonModule,
-    FieldsetModule,
-    InputTextModule,
-    PasswordModule,
+    MatButtonModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatInputModule,
     ReactiveFormsModule,
   ],
   declarations: [
