@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 
-import { Actions, Effect } from '@ngrx/effects';
+import { Actions, createEffect, ofType } from '@ngrx/effects';
 
-import { of } from 'rxjs/observable/of';
+import { of } from 'rxjs';
+import { catchError, mergeMap, switchMap } from 'rxjs/operators';
 
 import { GET_MENUBAR, ReceiveMenubarAction } from './menubar.actions';
 import { MenubarService } from './menubar.service';
@@ -12,23 +13,24 @@ import { RemoveTokenAction } from '../auth/auth.actions';
 @Injectable()
 export class MenubarEffects {
 
-  @Effect()
-  sendGetRequest$ = this.actions$
-    .ofType(GET_MENUBAR)
-    .switchMap(() => this.menubarService.get()
-      .mergeMap((response: any) => {
-        return [
-          new ReceiveMenubarAction(response.body),
-        ];
-      })
-      .catch(error => {
-        const unauthorizedCode = 401;
-        if (error.status === unauthorizedCode) {
-          return of(new RemoveTokenAction(''))
-        }
-        return of(new ReceiveMenubarAction(error));
-      })
-    );
+  
+  sendGetRequest$ = createEffect(() => this.actions$.pipe(
+    ofType(GET_MENUBAR),
+    switchMap(() => this.menubarService.get()
+      .pipe(
+        mergeMap((response: any) => {
+          return [
+            new ReceiveMenubarAction(response.body),
+          ];
+        }),
+        catchError(error => {
+          const unauthorizedCode = 401;
+          if (error.status === unauthorizedCode) {
+            return of(new RemoveTokenAction(''));
+          }
+          return of(new ReceiveMenubarAction(error));
+        }),
+      ))));
 
   constructor(private actions$: Actions,
               private menubarService: MenubarService,) {
