@@ -1,16 +1,23 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatToolbarModule } from '@angular/material/toolbar';
 
-import { MenubarModule } from 'primeng/primeng';
-
+import { MenubarComponent } from './menubar.component';
 import { MenubarContainer } from './menubar.container';
 import { MenubarService } from './menubar.service';
-import { MenubarComponent } from './menubar.component';
 
 @NgModule({
   imports: [
     CommonModule,
-    MenubarModule,
+    RouterModule,
+    MatButtonModule,
+    MatIconModule,
+    MatMenuModule,
+    MatToolbarModule,
   ],
   declarations: [
     MenubarComponent,
