@@ -2,27 +2,31 @@ import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { Location } from '@angular/common';
 
-import { Effect, Actions } from '@ngrx/effects';
+import { createEffect, Actions, ofType } from '@ngrx/effects';
+import { map, tap } from 'rxjs/operators';
 
 import { BACK, FORWARD, GO, Go } from './router.actions';
 
 
 @Injectable()
 export class RouterEffects {
-  @Effect({dispatch: false})
-  navigate$ = this.actions$.ofType(GO)
-    .map((action: Go) => action.payload)
-    .do(({path, query: queryParams, extras}) => {
+  
+  navigate$ = createEffect(() => this.actions$.pipe(
+    ofType(GO),
+    map((action: Go) => action.payload),
+    tap(({path, query: queryParams, extras}) => {
       return this.router.navigate(path, {queryParams, ...extras})
-    });
+    })), {dispatch: false});
 
-  @Effect({dispatch: false})
-  navigateBack$ = this.actions$.ofType(BACK)
-    .do(() => this.location.back());
+  
+  navigateBack$ = createEffect(() => this.actions$.pipe(
+    ofType(BACK),
+    tap(() => this.location.back())), {dispatch: false});
 
-  @Effect({dispatch: false})
-  navigateForward$ = this.actions$.ofType(FORWARD)
-    .do(() => this.location.forward());
+  
+  navigateForward$ = createEffect(() => this.actions$.pipe(
+    ofType(FORWARD),
+    tap(() => this.location.forward())), {dispatch: false});
 
   constructor(private actions$: Actions,
               private router: Router,
