@@ -1,14 +1,15 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
-import { AutoCompleteModule } from 'primeng/components/autocomplete/autocomplete';
-import { ButtonModule } from 'primeng/components/button/button';
-import { DataTableModule } from 'primeng/components/datatable/datatable';
-import { FieldsetModule } from 'primeng/components/fieldset/fieldset';
-import { InputTextModule } from 'primeng/components/inputtext/inputtext';
-import { MultiSelectModule } from 'primeng/components/multiselect/multiselect';
-import { PasswordModule } from 'primeng/components/password/password';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatIconModule } from '@angular/material/icon';
+import { MatInputModule } from '@angular/material/input';
+import { MatPaginatorModule } from '@angular/material/paginator';
+import { MatSortModule } from '@angular/material/sort';
+import { MatTableModule } from '@angular/material/table';
 
 import { TableContainer } from './table.container';
 import { TableComponent } from './table.component';
@@ -17,15 +18,17 @@ import { TableDataMappingComponent } from './table.data-mapping.component';
 
 @NgModule({
   imports: [
-    AutoCompleteModule,
-    ButtonModule,
     CommonModule,
-    DataTableModule,
-    FieldsetModule,
     FormsModule,
-    InputTextModule,
-    MultiSelectModule,
-    PasswordModule,
+    MatAutocompleteModule,
+    MatButtonModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
+    MatPaginatorModule,
+    MatSortModule,
+    MatTableModule,
     ReactiveFormsModule,
   ],
   declarations: [
