@@ -1,39 +1,62 @@
-import {$WebSocket} from 'angular2-websocket/angular2-websocket';
-import {Injectable} from '@angular/core';
-import {Store} from '@ngrx/store';
+import { Injectable, OnDestroy } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { Subject, fromEvent, takeUntil } from 'rxjs';
+import { map } from 'rxjs/operators';
 
-import {AppState} from '../app.reducers';
-import {AddRecordAction, DeselectRecordAction, RemoveRecordAction} from '../table/table.actions';
-
+import { AppState } from '../app.reducers';
+import { AddRecordAction, DeselectRecordAction, RemoveRecordAction } from '../table/table.actions';
 
 @Injectable()
-export class WebsocketService {
-  private ws: $WebSocket;
+export class WebsocketService implements OnDestroy {
+  private socket: WebSocket | null = null;
+  private destroy$ = new Subject<void>();
 
-  constructor(private store: Store<AppState>) {
-    // this.store.select(getRecords);
-  };
+  constructor(private store: Store<AppState>) {}
 
   connect(url: string) {
-    this.ws = new $WebSocket(url);
-    return this.ws.getDataStream()
-      .subscribe(
-        (response: any) => {
-          const message = JSON.parse(response.data);
-          switch (message.type) {
-            case 'INSERT':
-              this.store.dispatch(new AddRecordAction(message.row));
-              break;
-            case 'DELETE':
-              this.store.dispatch(new RemoveRecordAction(message.row));
-              this.store.dispatch(new DeselectRecordAction(message.row));
-              break;
-            default:
-              break;
+    if (this.socket) {
+      this.socket.close();
+    }
+    this.socket = new WebSocket(url);
+
+    fromEvent<MessageEvent>(this.socket, 'message')
+      .pipe(
+        takeUntil(this.destroy$),
+        map((event: MessageEvent) => {
+          try {
+            return JSON.parse(event.data);
+          } catch {
+            return null;
           }
-        },
-        (error: any) => console.log('Error: ' + error.message),
-        () => console.log('Completed'),
-      );
+        })
+      )
+      .subscribe(message => {
+        if (!message || !message.type) {
+          return;
+        }
+        switch (message.type) {
+          case 'INSERT':
+            this.store.dispatch(new AddRecordAction(message.row));
+            break;
+          case 'DELETE':
+            this.store.dispatch(new RemoveRecordAction(message.row));
+            this.store.dispatch(new DeselectRecordAction(message.row));
+            break;
+          default:
+            break;
+        }
+      });
+
+    fromEvent(this.socket, 'error')
+      .pipe(takeUntil(this.destroy$))
+      .subscribe();
+  }
+
+  ngOnDestroy() {
+    this.destroy$.next();
+    this.destroy$.complete();
+    if (this.socket) {
+      this.socket.close();
+    }
   }
 }
